refactor: websocket 패키지의 PickUpInfoRepository 삭제
* 기능을 pickup.repository.PickUpRepository로 통합
